backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.16,<4.0
django-cors-headers>=4.7,<5.0
psycopg[binary]>=3.2,<4.0
redis>=5.2,<6.0
celery>=5.4,<6.0
django-redis>=5.4,<6.0
